requirements.txt
Authlib==1.3.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-Bcrypt==1.0.1
idna==3.7
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pycparser==2.22
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
